I fixed the autodetection system to work on Mac, although I will be fixing that system up just a little bit
git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@437 4df4b0f3-56ce-47cb-b001-ed939b7d65a6

+++ src/sdl/CVideoDriver.h
@@ -12,7 +12,7 @@
 #define g_pVideoDriver CVideoDriver::Get()
 
 struct st_resolution
-{ short width,height,depth; };
+{ int width,height,depth; };
 
 #ifdef USE_OPENGL
 #include "COpenGL.h"
